+++ client/pages/Admin/Subscription/context.ts
@@ -1,9 +1,10 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 import { createContext } from 'react'
 import { SubscriptionSettings } from 'types'
 
 export interface ISubscriptionContext {
   settings: SubscriptionSettings
-  onSettingsChanged?: (key: string, value: any) => void
+  onChange?: (key: string, value: any) => void
 }
 
 export const SubscriptionContext = createContext<ISubscriptionContext>(null)