+++ server/services/mongo/user.ts
@@ -54,11 +54,14 @@ export class UserService extends MongoDocumentService<User> {
   /**
    * Get user by ID
    *
+   * @remarks Returns null if no user is found.
+   *
    * @param id - User ID
    */
-  public async getById(id: string) {
+  public async getById(id: string): Promise<User> {
     try {
       const user = await this.collection.findOne({ _id: id })
+      if (!user) return null
       if (!user.role) throw new Error(`The user ${id} has no role set.`)
       user.id = user._id
       user.role = await this._role.getByName(user.role as string)